server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.2
redis==5.0.7
celery[redis]==5.4.0
structlog==24.1.0
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==6.1.0
httpx==0.27.0
firebase-admin==6.5.0
pgvector==0.2.5
litellm==1.51.3
langgraph==0.2.24
openai==1.40.2
tenacity==9.0.0
typing-extensions==4.12.2
pytest==8.3.2
pytest-asyncio==0.23.8
sentry-sdk==2.12.0
opentelemetry-distro==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.27.0

